Fix circular dependencies and reorganize code
